+++ src/macpie/collections/base.py
@@ -5,15 +5,11 @@ from macpie.util.info import Info
 
 
 class BaseCollection(ABC):
-    """
-    Abstract base class for all collections.
+    """Abstract base class for all collections.
     """
 
     @abstractmethod
     def __iter__(self):
-        """
-            test
-        """
         while False:
             yield None
 
@@ -22,16 +18,14 @@ class BaseCollection(ABC):
         return {}
 
     def get_collection_info(self):
-        """
-        Contruct and return an :class:`macpie.util.Info` object
+        """Contruct and return an :class:`macpie.util.Info` object
         describing this collection.
         """
         collection_dict = {'class_name': self.__class__.__name__, **self.to_dict()}
         return Info.from_dict(collection_dict, title=get_option("sheet.name.collection_info"))
 
     def get_dataset_history_info(self):
-        """
-        Contruct and return an :class:`macpie.util.Info` object containing
+        """Contruct and return an :class:`macpie.util.Info` object containing
         all :attr:`macpie.Dataset.history` information.
         """
         info = Info(title=get_option("sheet.name.dsets_history"))